implemented medoid

+++ header_files/euclidean_distance.h
@@ -3,4 +3,4 @@
 #include <map>
 using namespace std;
 #pragma once
-void euclidean_distance(vector<vector<float>>&vec,vector<float>query_point, map<int,float>&);
+void euclidean_distance(vector<vector<float>>&vec,vector<float>query_point, map<int,double>&);
